Move informer_factory to staging
