+++ pkg/controller/resourcequota/BUILD
@@ -37,6 +37,7 @@ go_library(
         "//staging/src/k8s.io/client-go/listers/core/v1:go_default_library",
         "//staging/src/k8s.io/client-go/tools/cache:go_default_library",
         "//staging/src/k8s.io/client-go/util/workqueue:go_default_library",
+        "//staging/src/k8s.io/controller-manager/pkg/informerfactory:go_default_library",
         "//vendor/k8s.io/klog/v2:go_default_library",
     ],
 )
