+++ staging/src/k8s.io/controller-manager/pkg/informerfactory/informer_factory.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package informerfactory
 
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
