+++ test/integration/garbagecollector/BUILD
@@ -12,7 +12,6 @@ go_test(
     tags = ["integration"],
     deps = [
         "//cmd/kube-apiserver/app/testing:go_default_library",
-        "//pkg/controller:go_default_library",
         "//pkg/controller/garbagecollector:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1:go_default_library",
@@ -35,6 +34,7 @@ go_test(
         "//staging/src/k8s.io/client-go/metadata/metadatainformer:go_default_library",
         "//staging/src/k8s.io/client-go/restmapper:go_default_library",
         "//staging/src/k8s.io/client-go/tools/cache:go_default_library",
+        "//staging/src/k8s.io/controller-manager/pkg/informerfactory:go_default_library",
         "//test/integration:go_default_library",
         "//test/integration/framework:go_default_library",
     ],
